feat: modularize panels and types into separate modules
- Move panel logic, tools, types to src/modules/*
- Add fixed_panel_defaults() to Module trait for generic panel creation
- Generic ensure_default_contexts() using module registry
- Remove hardcoded types/enums/structs from state.rs (now module-owned)
- Migrate module-specific constants (deprecation timers, etc.) to modules
- Extract System prompts/seeds to dedicated system module (P0)
- Add needs_cache() trait method to Panel for background caching
- Update persistence to use HashMap<ContextType, UID> for important panels
- Re-export module types from state.rs for backwards compatibility
