+++ src/modules/core/tools/mod.rs
@@ -1,4 +1,3 @@
 pub mod close_context;
 pub mod message_status;
 pub mod manage_tools;
-pub mod system;
