+++ src/modules/glob/mod.rs
@@ -1,6 +1,9 @@
 mod panel;
 pub mod tools;
 
+/// Deprecation timer for glob panels (milliseconds)
+pub const GLOB_DEPRECATION_MS: u64 = 30_000; // 30 seconds
+
 use crate::core::panels::Panel;
 use crate::state::{ContextType, State};
 use crate::tool_defs::{ToolDefinition, ToolParam, ParamType, ToolCategory};
