+++ src/modules/system/panel.rs
