+++ src/modules/tree/panel.rs
@@ -11,6 +11,8 @@ use crate::ui::{theme, helpers::*};
 pub struct TreePanel;
 
 impl Panel for TreePanel {
+    fn needs_cache(&self) -> bool { true }
+
     fn handle_key(&self, key: &KeyEvent, _state: &State) -> Option<Action> {
         match key.code {
             KeyCode::Up => Some(Action::ScrollUp(SCROLL_ARROW_AMOUNT)),
