+++ src/tools/mod.rs
@@ -1,29 +1,10 @@
 use serde::{Deserialize, Serialize};
 use serde_json::Value;
 
-use crate::state::{estimate_tokens, ContextType, MessageStatus, State};
+use crate::state::State;
 
-/// Refresh token count for the Conversation context element
-pub fn refresh_conversation_context(state: &mut State) {
-    // Calculate total tokens from all active messages
-    let total_tokens: usize = state.messages.iter()
-        .filter(|m| m.status != MessageStatus::Deleted)
-        .map(|m| {
-            match m.status {
-                MessageStatus::Summarized => m.tl_dr_token_count.max(estimate_tokens(m.tl_dr.as_deref().unwrap_or(""))),
-                _ => m.content_token_count.max(estimate_tokens(&m.content)),
-            }
-        })
-        .sum();
-
-    // Update the Conversation context element's token count
-    for ctx in &mut state.context {
-        if ctx.context_type == ContextType::Conversation {
-            ctx.token_count = total_tokens;
-            break;
-        }
-    }
-}
+// Re-export from core module for backwards compatibility
+pub use crate::modules::core::conversation::refresh_conversation_context;
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct ToolUse {
